checked the endpoints for edge cases and handled exceptions for them

+++ src/main/java/com/gautami/authorization/controller/UserController.java
@@ -1,17 +1,28 @@
 package com.gautami.authorization.controller;
 
+import com.gautami.authorization.Repository.UserRepository;
 import com.gautami.authorization.dto.AdminRequest;
 import com.gautami.authorization.dto.UserDto;
+import com.gautami.authorization.exception.Forbidden;
 import com.gautami.authorization.exception.InvalidRequest;
+import com.gautami.authorization.exception.NotFoundException;
+import com.gautami.authorization.model.Role;
 import com.gautami.authorization.model.User;
 import com.gautami.authorization.service.UserService;
+import com.sun.security.auth.UserPrincipal;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
+import java.util.Set;
 
 @RestController
 @RequestMapping("/user")
@@ -24,6 +35,9 @@ public class UserController {
     @Autowired
     UserService userService;
 
+    @Autowired
+    UserRepository userRepository;
+
     @PostMapping("/register")
     @ResponseStatus(HttpStatus.CREATED)
     public void registerUser(@RequestBody UserDto userRequest) {
@@ -33,10 +47,10 @@ public class UserController {
 
     @PostMapping("/createadmin")
     @ResponseStatus(HttpStatus.CREATED)
-    public void createAdmin(@RequestBody AdminRequest request){
-        if(request.getKey().equals(secretKey)) {
+    public void createAdmin(@RequestBody AdminRequest request) {
+        if (request.getKey().equals(secretKey)) {
             userService.createAdminUser(request);
-        }else{
+        } else {
             //throw some error
             throw new InvalidRequest("The key given is not correct, please give correct key to proceed");
         }
@@ -52,7 +66,7 @@ public class UserController {
     @GetMapping("/id/{id}")
     @ResponseStatus(HttpStatus.OK)
     @PreAuthorize("hasRole('ADMIN')")
-    public User getUserById(@PathVariable Long id){
+    public User getUserById(@PathVariable Long id) {
         return userService.getUserById(id);
     }
 
@@ -60,17 +74,35 @@ public class UserController {
     @PutMapping("/update/{id}")
     @ResponseStatus(HttpStatus.OK)
     @PreAuthorize("hasAnyRole('ADMIN','USER')")
-    public void updateUser(@PathVariable Long id,@RequestBody UserDto userRequest){
-        userService.updateUser(id,userRequest);
+    public void updateUser(@PathVariable Long id, @RequestBody UserDto userRequest) {
+        userService.updateUser(id, userRequest);
     }
 
     @DeleteMapping("/delete/id/{id}")
     @ResponseStatus(HttpStatus.OK)
-    @PreAuthorize("hasAnyRole('ADMIN','USER')")
-    public void deleteUser(@PathVariable Long id){
-        userService.deleteUser(id);
+    @PreAuthorize("hasAnyRole('ADMIN', 'USER')")
+    public void deleteUser(@PathVariable Long id, @RequestParam(value = "userId") Long userId) {
+
+        Optional<User> permitUser = userRepository.findById(userId);
+        if (!permitUser.isPresent()) {
+            throw new NotFoundException("The user try to delete the account is not found");
+        }
+        Set<Role> permittedRole = permitUser.get().getRoles();
+
+        boolean flag = permittedRole.stream().anyMatch(role -> role.getRoleName().equalsIgnoreCase("ROLE_ADMIN"));
+        if (flag) {
+            userService.deleteUser(id);
+            return;
+        } else if (permitUser.get().getId() == id) {
+            userService.deleteUser(id);
+            return;
+        }
+        throw new Forbidden("You are not authorized to delete the user!!!!");
     }
 
 
-
 }
+
+
+
+
